requirements.txt
fastapi==0.103.2
uvicorn[standard]==0.23.2

# ORM stack — Pydantic v2 compatible
sqlmodel==0.0.21
sqlalchemy==2.0.25
alembic==1.12.1

# MySQL driver (you’re using mysql+mysqlconnector)
mysql-connector-python==8.1.0

# Auth & crypto
passlib==1.7.4
argon2-cffi==21.3.0
PyJWT==2.8.0
email-validator==2.1.0.post1

# Config & utils
python-dotenv==1.0.0
pydantic-settings==2.1.0

# ML
scikit-learn==1.3.2
joblib==1.3.2
lightgbm==4.1.0
sentence-transformers==2.2.2
huggingface-hub==0.23.4

# Testing
pytest==7.4.2
httpx==0.24.1

# Data utilities (seed/import helpers)
pandas==2.2.3
openpyxl==3.1.5
